Update categoryId, catalogId and etalaseId type to long

+++ tkpdseller/src/main/java/com/tokopedia/seller/product/data/repository/CatalogRepositoryImpl.java
@@ -21,7 +21,7 @@ public class CatalogRepositoryImpl implements CatalogRepository {
     }
 
     @Override
-    public Observable<CatalogDataModel> fetchCatalog(String keyword, int prodDeptId, int start, int row) {
+    public Observable<CatalogDataModel> fetchCatalog(String keyword, long prodDeptId, int start, int row) {
         return catalogDataSource.fetchCatalog(keyword, prodDeptId, start, row);
     }
 }
