+++ tkpdseller/src/main/java/com/tokopedia/seller/product/domain/interactor/FetchCatalogDataUseCase.java
@@ -27,13 +27,10 @@ public class FetchCatalogDataUseCase extends UseCase<CatalogDataModel>{
         this.catalogRepository = catalogRepository;
     }
 
-    public static RequestParams createRequestParams(String keyword,
-                                                    int departementId,
-                                                    int start,
-                                                    int rows){
+    public static RequestParams createRequestParams(String keyword, long departementId, int start, int rows){
         RequestParams params = RequestParams.create();
         params.putString("q", keyword);
-        params.putInt("sc", departementId);
+        params.putLong("sc", departementId);
         params.putInt("start", start);
         params.putInt("rows", rows);
 
@@ -43,7 +40,7 @@ public class FetchCatalogDataUseCase extends UseCase<CatalogDataModel>{
     @Override
     public Observable<CatalogDataModel> createObservable(RequestParams requestParams) {
         String q = requestParams.getString("q", "");
-        int sc = requestParams.getInt("sc", 0);
+        long sc = requestParams.getLong("sc", 0);
         int start = requestParams.getInt("start", 0);
         int rows = requestParams.getInt("rows", 20);
         return catalogRepository.fetchCatalog(q, sc, start, rows);
